+++ frontend/src/features/Labels/services/useLabelsQuery.tsx
@@ -1,0 +1,16 @@
+import { useQuery } from "@tanstack/react-query"
+import { LabelType } from "../../../interfaces";
+import { getLabels } from "./labelServices";
+
+
+export const useLabelsQuery = () => {
+
+  const {data, isSuccess, isPending, isError, error} = useQuery<LabelType[], Error>({
+    queryKey: ["labels"],
+    queryFn: () => getLabels()
+  })
+
+  return { data, isSuccess, isPending, isError, error };
+}
+
+export default useLabelsQuery
