+++ frontend/src/features/Notes/components/NoteComponents/HoverButtons.tsx
