+++ frontend/src/features/Notes/styles/MainStyles.module.css
