+++ frontend/src/features/Notes/styles/NoteStyles.module.css
